Only display "compile" scoped dependencies
Showing "compile" scoped dependencies to the user is good.  However,
we need to filter out all the other scopes.  MavenXpp3Reader doesn't use
"compile" scope by default, as maven does, so we have to do it.

We don't need to show "test" scoped dependencies.  They are mainly
useful to a developer of the project, who can just look at the
project.clj/pom.