Fix http uploads saving promoted file
Previously the http uploads would save the file and then check
if it could update the database. That would cause an error to be
returned since promoted artifacts should be immutable.  The
classic repo would then have a pom with a possible checksum error
and be out of sync with the releases repo.