Highlight org.clojars.* forks [issue #77]
This highlights org.clojars.* forks with italics anywhere jar-link is used,
which includes:

* the search results page
* the dashboard
* the projects browse page
* the header on the jar view page
* the header on the versions page (another change in this commit)
* probably other places I'm not aware of

It also adds a notice to the view page for a jar if it is a fork,
and a notice to the top of the search results if any forks are in
the result set. The notice links to the Groups wiki page for more
information.