Improve search result quality.
This should fix issues #197 and #194.

We no longer give so much weight to the number of downloads.
Previously, the number of downloads could change an item's search
score by a factor of about 500,000x (for lein-ring, with approx. 500K
downloads).  This meant that a search that had any terms in common
with popular items almost always returned those items at the top, even
if there was an item that had a perfect text match.

With this change, lein-ring's downloads would only give it a
multiplier of about 2.43x.