Replace the ring-servlet monkey patch with configuator
It is difficult to determine where to have this kind of monkey
patching occur. It needs to happen for starting the production system,
starting any repl based development system, and for the integration tests.
Additionally, this kind of monkey patching is brittle if ring-jetty
changes it's implementation details. This patch removes it.

As an alternative the http server is configured with a decorator
around the HttpServletResponse. It checks for a custom header
and when it is found, it sets the status message. This could be slower
since it has a check everytime a header is set. An alternative
is to decorate the getOutputStream and getWriter with custom ones that
check on .flush and set the status beforehand.