Thread the database through the web app
Now that korma is removed each db call creates a new connection.
Step one of using a connection pool is to thread the db through
the web app as a parameter.

The tests see significant changes to account for the new apis.
Additionally the fixures change to allow keeping one connection
to the database per test, and an in memory database can now be used.
All together this reduces test runtime by 10 seconds on my machine.

The clojars admin namespace contains a dynamic var for its database
connection.  The nrepl server uses some middleware to populate that
var for connections.

The one off tasks (migrations,setup) are changed to use the new apis.

Unfortunatly `lein ring` no longer works as a database needs to be
passed to the handler. I plan for a reloaded repl based workflow to
replace the dev server functionality.