Replace korma with yesql
DB access drops down to sql in serveral places. Korma does not appear to
be providing a good abstraction. This commit removes korma and replaces
it with yesql. This allows sql to be in sql without the string
concatenation that raw clojure.jdbc would require.

The database functions will read the config each time. This is
unfortunate, but allows a dev and test config to be used in the
same jvm/repl.  The next step will be to thread the db connection through
the system to allow a connection pool to be used.