Move yeller into a component and protocolize api
Several places were using error reporting through yeller. This patch
threads an ErrorReporter though the app. A protocol is used to allow
injection of specific implementations for each system. The production
system gets an implementation that prints to stdout and reports to
yeller. The development system gets an implementation that only prints
to stdout.  Testing gets an implementation that ignores errors, as the
resulting test failure should provide context.