Move database functions into a protocol
Duct recommends using protocols around your components, so lets
make one for the database.  This allows the thread executor to stop
being a global. However we need to make the database functions into a
protocol.  This will allow future work on having a quick/fake database
for testing if desired.