Thread object for stats lookup and use a protocol
This allows different implementations based on usage:

1. FileStat for dev and production system. This one reads a
file (memoized!) just like previously. It takes a filesystem factory,
so that unit tests can be created for it that use an in memory filesystem.

2. MapStat for in memory version. This is delegated to for computation
by FileStat. It is also used in the tests to provide 0 downloads without
having to touch a filesystem.

3. A reified stub used for unit testing search.