Thread object for search and use a protocol
Create LuceneSearch for production/development and integration testing.
Allow use of an index-factory for an in memory index for tests. A
no-search one is used for tests that don't need it.

Since one of the integration tests adds data to the index, and then
searches the started system is now saved and accessible. This might be
better as a unit test later once the db is abstracted, and unit testing
the handlers is possbile.

Also add some unit tests for LuceneSearch. This will make it easier to build
other ones for any of the search bugs/feature requests.