Thread mailer through system
Pass a function through as a mailer. Use this for sending password
resets.

The prod/dev systems can use the normal one. Tests are added for it
which spawn an smtp server and connect to it.  Tests for other parts of
the system can pass a mailer that just sends the data to a promise for
verification.