Don't use stop words when indexing for search.
This disables the default list of stop words used by Lucene's
StandardAnalyzer. Stop words don't work well when indexing text that
tends more toward non-prose.

In my tests, this increases the size of the Lucene index on disk by
about 1%, but gives better search results.  For example, now a search
for "at" will return an item that has the artifact-id "at-at" or an item
that references "at-at" in its description.

Fixes #243.