Add tool to repair broken maven-metadata [#455]
Without atomic deploys, we end up with deploys that don't complete
successfully, so leave maven-metadata.xml files in a state that is
inconsistent with the actually deployed versions. Once the metadata gets
in this state, the problem worsens, since aether will deploy the pom and
jar files, then will fail to update the metadata file if its sum isn't
correct.

This adds a command-line tool that can scan the repo to report on or
repair metadata files that has a version list that doesn't match the
actual versions, or has invalid/missing sums.

This tool should only be needed until atomic deploys are implemented,
which will prevent the inconsistencies from occurring.