Don't ignore files uploaded after maven-metadata.xml [#515 #532]
Maven will upload *any* classified artifact after uploading
maven-metadata.xml. We use maven-metadata.xml as a trigger to know when
to finalize the deploy, moving the contents to the actual repo and to
cloudfiles. Before this change, we essentially ignored any files that
came in after the finalization. Now, any files that come in after
finalization go straight to the actual repo and to cloudfiles.

This means that deployments with classified artifacts is no longer
atomic - the primary artifact, pom, and maven-metadata.xml will be
consistent, but any classified artifacts won't be part of that atomic
unit.

This commit also switches us from clj-http-lite to clj-http for
testing, since we need cookie support to properly test this change,
since we need to control the order of deployment, so can't use aether.