Improve pom import
This allows the index of poms to be imported to be generated ahead of
time, reducing downtime for the site when we are applying it.

It also uses the path to the pom file as the canonical id of the
artifact instead of the values in the pom, since we didn't always verify
they were the same, and have some poms that lie.

Lastly, it now outputs status as it goes, to allow gauging how long it
will take.