Don't share tmp repos between artifacts [#541]
lein-module based projects will share an aether session when deploying,
but we want each module to have it's own session, so they will each have
their own atomic deploy. Without this change, the second and all
subsequent module deploys will fail.

The fix here is to store the tmp repos as a seq, and find the existing
one that matches the current upload, or creating a new one if none
match. This would allow interleaving files for different artifacts in
the same session, which none of the existing tooling does, but who knows
what craziness exists out there?