Clean up dependencies
This removes all pedantic warnings, and sets the build to abort if new
warnings are introduced.

It also removes some dependencies that we get transitively, and updates
some versions to aid in removing the warnings.

The tests caught one behavior change in the status code we return for a
PUT to a missing route. This should be fine, as users shouldn't be
depending on that error code.