Be forgiving of extra space after date in logs
With the recent nginx update, we changed the access log format slightly:

* 2 spaces after the date instead of 1
* append the sni host to the end of the line

The log parser ignores the sni host, but was strict about the spaces
after the date, causing stat generation to find no downloads. This
relaxes the regex to allow for any number of spaces after the date so we
can re-process logs that were generated with the new format, and allows
us to adjust the format to go back to one space if we decide to do so.

Related to #554