Suggest looking in maven central for some artifacts
For a list of artifacts we know live in maven central
(org.clojure ones) we suggest going to maven central for
them.

This is pretty rough at the moment, still needs some styling
(imo) and we should probably move the artifacts out to config
somewhere rather than keeping them in code.

Also cleaned up some of the styling on the message flash