Add stats processing for cdn access logs
Fastly gives us a log file for every minute for every endpoint that
processed requests, so we first combine all of those logs into one file
for yesterday, and stash a copy of that in cloudfiles. We then combine
that file with the local nginx access log and process stats.

Since the logging format from fastly isn't CLF, we need to treat it a
bit differently when parsing.

This also includes a script for updating stats for any day, allowing us
to merge in the cdn stats we collected before this goes live.