Don't allow deploys where group-id/artifact-id exists on central
This prevents shadowing of libraries released to maven central. It
checks by seeing if it can access the maven-metadata.xml from central,
and does an exponential backoff for up to 512ms if central can't be
contacted. If it completely fails to contact central, it fails the
deploy to prevent opportunistic shadow deploys during rare periods when
central is down.

Related to #468