Support sharing session between deploys [#637]
We currently store the tmp upload dirs in the session, and look up the
correct one by matching group and artifact name. If multiple SNAPSHOT
deploys share the same session and group/artifact, latter deploys
will fail because they will attempt to reuse the former finalized
upload dir.

This modifies the upload dir checking logic to take into account the
version and the SNAPSHOT's timestamped version when determining if the
dir should be reused for the current upload. This also modifies the
metadata writing logic to not overwrite values derived from artifacts
uploaded previously within the same deploy, which could happen depending
on the upload order.