Allow whitelisting artifacts that can shadow central
This uses an EDN file in the build itself - we'll update this so
infrequently that I don't think it needs all the work to store it in the
DB.

The artifact triggering this change (net.mikera/clojure-pom) was
deployed to central in the past, but has been exclusively deployed to
clojars for the last few years.

Fixes #639.