Allow maven-metadata.xml to be uploaded twice
Maven now (as of 3.5.0, maybe earlier) uploads a second, identical
maven-metadata.xml if you are uploading artifacts with
classifiers (-test, -sources, etc). This breaks the
validation/finalization process, since we use that file as a trigger to
finalize. When it is uploaded a second time, we revalidate, which will
fail because validation falsely determines you are trying to overwrite
an existing deploy.

The fix here is to skip the finalization step if we've already finalized
AND the maven-metadata.xml file is identical to the one we already have
in the upload dir.

Fixes #640.