Upload to cloudfiles synchronously
This allows us to fail the build if we can't upload to cloudfiles. We
are moving towards the blobstore being the canonical repo, so want to
make sure we write to it successfully before returning success to the
user.

This removes the async process for uploading the artifacts, and
removes durable-queue from the dependencies since that was the only
functionality that was using it.

This is related to [#707]