Upload artifacts to s3 on deploy
This uploads a copy of the deployed artifacts to s3 in addition to
clouldfiles. This allows for an easier transition - once the existing
repo is copied over and this has been deployed, we can switch the CDN
to point to the s3 bucket any time without downtime.

This also includes a rewrite of the s3 protocol to encapsulate the
bucket name. This allows for less configuration to be passed around,
but does require an s3 client per bucket.