Handle XSS edge case with edn in db
We store :licenses and :scm information for a jar as edn maps in the
db. This data is generated by our own code when reading poms, so we
should have full control over it. But, if someone were able to find a
way to get hiccup data into that edn data, it could allow an XSS
exploit since our escaping code escapes html from strings, but not
from hiccup data.

This adds validation to ensure that the data is the correct shape both
going in to and out of the db. We will throw if there is an attempt to
insert invalid data, but will treat invalid data on read as a nil to
prevent it from interrupting normal operation.