Add hash values to deploy tokens
This will allow us to lookup tokens by value in a reasonable amount of
time when handling a breach notification from GitHub. Before this
change, we would read *all* tokens from the db, then bcrypt-verify
each one until we found the token. That process took ~500ms for *each*
token, so wasn't really sustainable.

The new approach writes a one-way sha256 hash of the token to the db
when it is created, then we look up all the tokens that have that
hash (will be a set of 1 in almost all cases), then we bcrypt-verify
all tokens in that set until we find it.

We have tokens in the db already that don't have hashes, so we write
the hash to the db on token use if one isn't already set.

This means that we can't detect breaches for tokens that were created
before this point and haven't been used since this point to write the
hash.