Allow login with GitHub
It just uses the e-mail on the users table to validate
the user. It also verifies if the user's e-mail is verified
by GitHub, so it would not allow anyone to just create a fake
account with the same e-mail.

I used ScribeJava, because is well maintained and the Clojure
alternative is not updated for some time. I just built an abstraction
using a `SocialProcotol`, which can be used in the future to
build other options.