Handle errors returned by the OAuth provider
When GitHub returns an error to the callback
handle it or throw an exception. Since it would
be interesting for the Clojars administration to
know when errors like `application_suspended` or
`redirect_uri_mismatch` happen, it throws an
exception for these cases.

For the `access_denied` error it inform the user
with an "You declined access to your account"
error.