Verify github groups on login
This will create and verify com/io.github.<gh-username> groups when a
user logs in via GitHub. It will only do so if the group isn't
verified, and will only create if the group doesn't exist.

If the group exists but the logging in user doesn't have access to it,
we just log and do nothing.

Fixes #784