Remove verified badge for non-owned groups
Showing the verified badge publicly for groups places a stigma on
legacy groups that can't be verified, and implies a false sense of
security for groups that are verified.

This removes the verified badge from the public view, only showing it
to a logged in user when they are viewing their own groups on the
dashboard or group page.