Allow deploying using email address
We document that the username is required to deploy, but allow users to
log in via their email address, which leads some users to think that
their email address is their username.

This expands deployment to support using the email address to match what
we allow at login.

This fixes #620.