Fix adding/toggling member to group
There was a bug in the group management code that prevented adding a
user to a group as a member. Any attempt resulted in adding them as an
admin, even though the UI would tell you they are a member on the first
page load after adding them (future page loads would show them as
admins). Attempts to toggle admin off will appeared to work, but the
user would remain an admin.

Fixes #838