Require licenses for new projects
This will require that any project that is new to have a license in the
pom for every released version.

This allows existing projects (ones that had a version released before
the boundary date) to continue to deploy versions without licenses, but
only if the most recent release did not have a license. If the project
has a license, it must continue to provide one. We will remove that in
the future after giving ample notice, at which point every deploy must
have a license.