Directly depend on slf4j-api
This is a bit of a mystery: if we let slf4j-api come in via a transitive
dependency, our logback.xml file is ignored (at least in dev & test), so
all logging goes to stdout. clojure.tools.logging is finding slj4j-api
in both cases, but logback does not get initialized in the case where
this isn't a top-level dep.

I have no idea why this is happening! Logging is terrible.