Properly delete artifacts from s3
We have admin functionality to delete artifacts from the repo that we
use rarely. This code had a bug where it would not delete from the s3
repo. This was because the s3-client assumes that a string path that
does not end in / is a single object, and the presence of the / means we
need to delete multiple objects. The admin delete code was generating a
path that did not end in /, but assumed that the s3-client would delete
everything under that path.

This fixes that issue by generating a path with a trailing /, and
expands the admin deletion test to test against s3.

Fixes #881.