Constrain memory used by cron-run scripts
By default, the JVM will use a heap of 960GB on the current instance
size we are using if Xmx isn't provided. This ensures that the scripts
will use only 512GB, and will fully allocate it on startup to fail-fast
if it isn't available.

This also sets the jobs to exit if they OOM, and dump a heap in that
case.

This is to help prevent memory contention with the server process.