[search] Rework search to always use db jar data
Search was originally pom-driven, then grew to support indexing either
pom data or db jar data. We only used the pom data when indexing on
upload, and ended up with a different :at timestamp value, depending on
the source of the data.

This simplifies the indexing to only support db jar data as input, which
fixes that discrepancy.

Fixes #897.