Store partial uploads in local tmp
This moves the uploads from /tmp to ~/tmp, as the former is a small
partition (3.8GiB) and the latter is on the root partition (40GiB). This
helps reduce the chance that lots of uploads will consume all the disk
space before the cron job to remove old uploads can clear it out.