Don't throw with invalid address on password reset
If the user doesn't enter a value on the password reset form, we look up
a user with an email of "", which matches deleted users. We then try to
send an email to that address, which throws.

This now catches and logs that case, and gives the default message to
the user that we will send an email if a user was found.